# docs/swagger.yaml
basePath: /api/v1
definitions:
  dsl.Arg:
    properties:
      i:
        type: integer
      m:
        additionalProperties:
          type: string
        type: object
      s:
        type: string
    type: object
  dsl.Op:
    properties:
      args:
        items:
          $ref: '#/definitions/dsl.Arg'
        type: array
      name:
        type: string
    type: object
  dto.AssignmentHistoryResponse:
    properties:
      cluster_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      matched_pattern:
        type: string
      raw_captures:
        additionalProperties:
          type: string
        type: object
      rendered_cluster_key:
        type: string
      rule_id:
        type: string
      transformed_captures:
        additionalProperties:
          type: string
        type: object
      trigger_source:
        type: string
      warnings:
        type: string
    type: object
  dto.BulkRedistributeRequest:
    properties:
      element_ids:
        items:
          type: string
        type: array
    type: object
...
  dto.ElementResponse:
    properties:
      assignment_status:
        type: string
      cluster_id:
        type: string
      first_seen_at:
        type: string
      id:
        type: string
      ip_address:
        type: string
      last_rule_id:
        type: string
      last_seen_at:
        type: string
      sys_name:
        type: string
    type: object
  dto.ReorderRequest:
    properties:
      rule_ids:
        items:
          type: string
        type: array
    type: object
  dto.RuleRequest:
    properties:
      capture_transforms:
        additionalProperties:
          items:
            $ref: '#/definitions/dsl.Op'
          type: array
        type: object
      cluster_key_template:
        type: string
      description:
        type: string
      is_enabled:
        type: boolean
      name:
        type: string
      pattern:
        type: string
      priority:
        type: integer
    type: object
  dto.RuleResponse:
    properties:
      capture_transforms:
        additionalProperties:
          items:
            $ref: '#/definitions/dsl.Op'
          type: array
        type: object
      cluster_key_template:
        type: string
      description:
        type: string
      id:
        type: string
      is_enabled:
        type: boolean
      name:
        type: string
      pattern:
        type: string
      priority:
        type: integer
    type: object
  dto.RuleTestRequest:
    properties:
      capture_transforms:
        additionalProperties:
          items:
            $ref: '#/definitions/dsl.Op'
          type: array
        type: object
      cluster_key_template:
        type: string
      hostname:
        type: string
      pattern:
        type: string
    type: object
  dto.RuleTestResponse:
    properties:
      matched:
        type: boolean
      raw_captures:
        additionalProperties:
          type: string
        type: object
      rendered_cluster_key:
        type: string
      transformed_captures:
        additionalProperties:
          type: string
        type: object
      warnings:
        items:
          type: string
        type: array
      would_create_new_cluster:
        type: boolean
    type: object
info:
  contact: {}
  description: SNMP discovery + rule-based virtual cluster assignment for network
    devices (Phase 1).
  title: Netcluster API
  version: "1.0"
paths:
...
  /elements:
    get:
      parameters:
      - description: Filter by cluster ID
        in: query
        name: cluster_id
        type: string
      - description: Filter by assignment status (unassigned|assigned)
        in: query
        name: assignment_status
        type: string
      - description: Free-text search
        in: query
        name: q
        type: string
      - description: Max rows to return (default 100, max 1000)
        in: query
        name: limit
        type: integer
      - description: Rows to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ElementResponse'
            type: array
      summary: List network elements
      tags:
      - elements
  /elements/{id}:
    delete:
      parameters:
      - description: Element ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Soft-delete a network element
      tags:
      - elements
    get:
      parameters:
      - description: Element ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ElementResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get a network element by ID
      tags:
      - elements
  /elements/{id}/history:
    get:
      parameters:
      - description: Element ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.AssignmentHistoryResponse'
            type: array
      summary: Get an element's assignment history timeline
      tags:
      - elements
  /elements/{id}/redistribute:
    post:
      parameters:
      - description: Element ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Re-run the rule engine for one element
      tags:
      - elements
  /elements/redistribute:
    post:
      consumes:
      - application/json
      parameters:
      - description: Optional element_ids to scope to
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.BulkRedistributeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Re-run the rule engine for a set of elements (or all, if none given)
      tags:
      - elements
  /rules:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.RuleResponse'
            type: array
      summary: List rules ordered by priority
      tags:
      - rules
    post:
      consumes:
      - application/json
      parameters:
      - description: Rule definition
        in: body
        name: rule
        required: true
        schema:
          $ref: '#/definitions/dto.RuleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.RuleResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Create a rule
      tags:
      - rules
  /rules/{id}:
    delete:
      parameters:
      - description: Rule ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Delete a rule
      tags:
      - rules
    get:
      parameters:
      - description: Rule ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RuleResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get a rule by ID
      tags:
      - rules
    put:
      consumes:
      - application/json
      parameters:
      - description: Rule ID
        in: path
        name: id
        required: true
        type: string
      - description: Rule definition
        in: body
        name: rule
        required: true
        schema:
          $ref: '#/definitions/dto.RuleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RuleResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Update a rule
      tags:
      - rules
  /rules/reorder:
    post:
      consumes:
      - application/json
      parameters:
      - description: Ordered rule IDs
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ReorderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Reorder rule priorities
      tags:
      - rules
  /rules/test:
    post:
      consumes:
      - application/json
      description: 'Evaluates a candidate rule (not necessarily persisted) against
        a sample hostname and reports what would happen, without writing anything:
        no element, no cluster, no history row.'
      parameters:
      - description: Candidate rule + sample hostname
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.RuleTestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RuleTestResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Dry-run a candidate rule against a sample hostname
      tags:
      - rules
swagger: "2.0"